requirements.txt
# Core dependencies
smolagents>=0.1.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
scipy>=1.11.0
statsmodels>=0.14.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Configuration and validation
pydantic>=2.0.0
pyyaml>=6.0.0
python-dotenv>=1.0.0

# Data processing
chardet>=5.0.0
openpyxl>=3.1.0
pyarrow>=12.0.0

# LLM providers
openai>=1.0.0
ollama>=0.1.0
google-generativeai>=0.3.0

# Reporting
python-pptx>=0.6.21
reportlab>=4.0.0
jinja2>=3.1.0

# CLI and user interface
click>=8.1.0
rich>=13.0.0

# HTTP and networking
requests>=2.31.0
httpx>=0.28.0

# Utilities
duckduckgo-search>=8.0.0
pathlib2>=2.3.0

# Development and testing (optional)
pytest>=7.0.0
pytest-cov>=4.0.0
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0
mypy>=1.0.0
pre-commit>=3.0.0